WIP refactoring: this is really dirty (and would be cleaned later)
A new Driver class has been created. The Driver class is in charge of
creating the network, creating machine and running the test script.
A Driver takes as parameter the class of the Machine it creates. It is
then possible to define specialized Machine subclasses.

Several methods has been added to the Machine class:
- `log_line(line)`: this method is called on each line outputed by the
  machine
- `cleanup`: called by the driver on exit

A driver script for Tmux has been added. It defines the class
`MachineTmux` which runs machines in Tmux windows.